LayoutTests/platform/chromium/http/tests/security/xss-eval-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/xss-eval3.html from frame with URL http://127.0.0.1:8000/security/resources/xss-eval2.html. Domains, protocols and ports must match.

This page verifies that you can't use eval to subvert cross-domain checks.

If the test passes, you'll see a pass message below.

PASS: documentFromEval should be It's me! and is.
PASS: childEval.call(frames[0], 'document') should be EvalError and is.
PASS: childEvalCaller('document') should be TypeError and is.
PASS: childLocalEvalCaller('document') should be EvalError and is.
